backend/pyproject.toml
[project]
name = "gemini-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.3",
    "google-genai>=1.62.0",
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.40.0",
]
